pyproject.toml
[project]
name = "diarization-transcription"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "diart>=0.9.1",
    "huggingface-hub>=0.27.0",
    "numpy==1.24.4",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
kafka = [
    "kafka-python>=2.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
